/ ‘ledger-sandbox.py’

+++ ledger-sandbox.py
@@ -27,6 +27,6 @@ for transaction in transactions:
     
     for line in transaction_lines[1:]:
         if (not transaction_dict['total']) and (len(line.rsplit('  ',1)) > 1):
-            transaction_dict['total'] = int(line.rsplit('  ',1)[-1].strip('$'))
+            transaction_dict['total'] = float(line.rsplit('  ',1)[-1].strip('$'))
             
     print transaction_dict
